src/telegram-bot/requirements.txt
pyTelegramBotAPI==4.18.1
aiohttp==3.9.5
python-dotenv
